Refactor: agregar validaciones, usar Integer y reorganizar estructura

+++ src/main/java/Metaphorce/Assessment2/GestionDeTareas/Service/TaskService.java
@@ -1,5 +1,7 @@
-package Metaphorce.Assestment2.Gestion.de.Tareas;
+package Metaphorce.Assessment2.GestionDeTareas.Service;
 
+import Metaphorce.Assessment2.GestionDeTareas.Entity.Task;
+import Metaphorce.Assessment2.GestionDeTareas.Repository.TaskRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import java.util.List;
@@ -9,7 +11,7 @@ import java.util.Optional;
 public class TaskService {
 
     @Autowired
-    private TaskRepository taskRepository;
+    TaskRepository taskRepository;
 
     public Task createTask(Task task) {
         return taskRepository.save(task);
@@ -19,23 +21,26 @@ public class TaskService {
         return taskRepository.findAll();
     }
 
-    public Optional<Task> getTaskById(Long id) {
-        return taskRepository.findById(id);
+    public Task updateTask(Integer id, Task updatedTask) {
+        Optional<Task> taskOptional = taskRepository.findById((id));
+
+        if (taskOptional.isPresent()) {
+            Task t = taskOptional.get();
+            t.setCompleted(true);
+            return taskRepository.save(t);
+        } else {
+            throw new RuntimeException("Task not found with ID: " + id);
+        } // Manejo de error con una estructura distinta
     }
 
-    public Task updateTask(Long id, Task updatedTask) {
-        return taskRepository.findById(id).map(task -> {
-            task.setTitle(updatedTask.getTitle());
-            task.setCompleted(updatedTask.isCompleted());
-            return taskRepository.save(task);
-        }).orElseThrow(() -> new RuntimeException("Task not found with id " + id)); // Manejo básico de error
-    }
-
-    public void deleteTask(Long id) {
+    public void deleteTask(Integer id) {
+        if (!taskRepository.existsById(id)){
+            throw new RuntimeException("Task not found with ID: " + id);
+        }
         taskRepository.deleteById(id);
     }
 
     public List<Task> getUncompletedTasks() {
-        return taskRepository.findByCompletedFalse();
+        return taskRepository.findByCompleted(false);
     }
 }
